api/requirements.txt
aiohttp==3.11.11
asyncpg==0.30.0
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.37
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.2.1
pydantic-settings==2.7.1
